requirements.txt
django==4.1.1
django-simple-captcha==0.5.17
django-crontab~=0.7.1
qrcode==7.3.1
numpy==1.23.3
pandas==1.5.0
matplotlib==3.6.0
reportlab~=3.6.11
svglib~=1.4.1
python-dotenv~=0.21.0
psycopg2~=2.9.3
seaborn~=0.12.0
six~=1.16.0
dnspython~=2.3.0
xlsxwriter~=3.0.9
Pillow==9.5.0